% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundles.R
\name{apply_bundle_filter}
\alias{apply_bundle_filter}
\title{Apply bundle filter to assets query}
\usage{
apply_bundle_filter(tbl, bundle_def)
}
\arguments{
\item{tbl}{Assets table (dbplyr tbl)}

\item{bundle_def}{Bundle definition list (the "where" clause)}
}
\value{
Filtered table
}
\description{
Applies the bundle's where clause as dplyr filters on the assets table.
}
\keyword{internal}
