% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R, R/backends_aws.R,
%   R/backends_balsa.R, R/backends_local.R, R/backends_rest.R
\name{backend_download}
\alias{backend_download}
\alias{backend_download.hcpx_backend_aws}
\alias{backend_download.hcpx_backend_balsa}
\alias{backend_download.hcpx_backend_local}
\alias{backend_download.hcpx_backend_rest}
\title{Download a file}
\usage{
backend_download(backend, url_or_path, dest, resume = TRUE, ...)

\method{backend_download}{hcpx_backend_aws}(backend, url_or_path, dest, resume = TRUE, ...)

\method{backend_download}{hcpx_backend_balsa}(backend, url_or_path, dest, resume = TRUE, ...)

\method{backend_download}{hcpx_backend_local}(backend, url_or_path, dest, resume = TRUE, ...)

\method{backend_download}{hcpx_backend_rest}(backend, url_or_path, dest, resume = TRUE, ...)
}
\arguments{
\item{backend}{Backend object}

\item{url_or_path}{URL or path to download}

\item{dest}{Local destination path}

\item{resume}{Whether to resume partial downloads}

\item{...}{Additional arguments}
}
\value{
Tibble with status, bytes
}
\description{
Download a file

Download file via AWS backend

Download file via local backend

Download file via REST backend
}
\keyword{internal}
