% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R, R/backends_aws.R,
%   R/backends_balsa.R, R/backends_local.R, R/backends_rest.R
\name{backend_head}
\alias{backend_head}
\alias{backend_head.hcpx_backend_aws}
\alias{backend_head.hcpx_backend_balsa}
\alias{backend_head.hcpx_backend_local}
\alias{backend_head.hcpx_backend_rest}
\title{Get file metadata via HEAD request}
\usage{
backend_head(backend, remote_path, ...)

\method{backend_head}{hcpx_backend_aws}(backend, remote_path, ...)

\method{backend_head}{hcpx_backend_balsa}(backend, remote_path, ...)

\method{backend_head}{hcpx_backend_local}(backend, remote_path, ...)

\method{backend_head}{hcpx_backend_rest}(backend, remote_path, ...)
}
\arguments{
\item{backend}{Backend object}

\item{remote_path}{Remote file path}

\item{...}{Additional arguments}
}
\value{
Tibble with size_bytes, etag, last_modified
}
\description{
Get file metadata via HEAD request

Get file metadata via AWS backend

Get file metadata via local backend

Get file metadata via REST backend
}
\keyword{internal}
