% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R, R/backends_aws.R,
%   R/backends_balsa.R, R/backends_local.R, R/backends_rest.R
\name{backend_presign}
\alias{backend_presign}
\alias{backend_presign.hcpx_backend_aws}
\alias{backend_presign.hcpx_backend_balsa}
\alias{backend_presign.hcpx_backend_local}
\alias{backend_presign.hcpx_backend_rest}
\title{Generate presigned URL for download}
\usage{
backend_presign(
  backend,
  remote_path,
  expires_sec = .DEFAULT_PRESIGN_EXPIRY_SEC
)

\method{backend_presign}{hcpx_backend_aws}(
  backend,
  remote_path,
  expires_sec = .DEFAULT_PRESIGN_EXPIRY_SEC
)

\method{backend_presign}{hcpx_backend_balsa}(
  backend,
  remote_path,
  expires_sec = .DEFAULT_PRESIGN_EXPIRY_SEC
)

\method{backend_presign}{hcpx_backend_local}(
  backend,
  remote_path,
  expires_sec = .DEFAULT_PRESIGN_EXPIRY_SEC
)

\method{backend_presign}{hcpx_backend_rest}(
  backend,
  remote_path,
  expires_sec = .DEFAULT_PRESIGN_EXPIRY_SEC
)
}
\arguments{
\item{backend}{Backend object}

\item{remote_path}{Remote file path}

\item{expires_sec}{Seconds until URL expires}
}
\value{
Character URL
}
\description{
Generate presigned URL for download

Generate presigned URL via AWS backend

Generate presigned URL via local backend

Generate presigned URL via REST backend
}
\keyword{internal}
