% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_ensure_dir}
\alias{cache_ensure_dir}
\title{Ensure cache directory exists for a path}
\usage{
cache_ensure_dir(path)
}
\arguments{
\item{path}{File path}
}
\value{
Invisible path
}
\description{
Ensure cache directory exists for a path
}
\keyword{internal}
