% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_list}
\alias{cache_list}
\title{List cached files}
\usage{
cache_list(h, pinned_only = FALSE)
}
\arguments{
\item{h}{hcpx handle}

\item{pinned_only}{If TRUE, only show pinned files}
}
\value{
Tibble of cached files
}
\description{
List cached files
}
\examples{
\dontrun{
h <- hcpx_ya()
cache_list(h)
}
}
