% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_path_from_asset}
\alias{cache_path_from_asset}
\title{Get cache path from asset row}
\usage{
cache_path_from_asset(h, asset)
}
\arguments{
\item{h}{hcpx handle}

\item{asset}{Tibble row with release, subject_id, remote_path columns}
}
\value{
Local file path
}
\description{
Get cache path from asset row
}
\keyword{internal}
