% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{canonical_task_names}
\alias{canonical_task_names}
\title{List all canonical task names}
\usage{
canonical_task_names()
}
\value{
Character vector of canonical task names
}
\description{
List all canonical task names
}
\keyword{internal}
