% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_seed.R
\name{catalog_clear}
\alias{catalog_clear}
\title{Clear all catalog data (for testing)}
\usage{
catalog_clear(h)
}
\arguments{
\item{h}{hcpx handle}
}
\value{
Invisible handle
}
\description{
Clear all catalog data (for testing)
}
\keyword{internal}
