% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_connect.R
\name{catalog_connection_valid}
\alias{catalog_connection_valid}
\title{Check if a database connection is valid}
\usage{
catalog_connection_valid(con)
}
\arguments{
\item{con}{DBI connection}
}
\value{
Logical TRUE if connection is valid
}
\description{
Check if a database connection is valid
}
\keyword{internal}
