% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_connect.R
\name{catalog_db_path}
\alias{catalog_db_path}
\title{Get database file path from connection}
\usage{
catalog_db_path(con)
}
\arguments{
\item{con}{DBI connection}
}
\value{
Character path to database file
}
\description{
Get database file path from connection
}
\keyword{internal}
