% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_connect.R
\name{catalog_disconnect}
\alias{catalog_disconnect}
\title{Disconnect from catalog database}
\usage{
catalog_disconnect(con)
}
\arguments{
\item{con}{DBI connection}
}
\value{
Invisible NULL
}
\description{
Disconnect from catalog database
}
\keyword{internal}
