% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_seed.R
\name{catalog_has_data}
\alias{catalog_has_data}
\title{Check if catalog has seed data}
\usage{
catalog_has_data(h)
}
\arguments{
\item{h}{hcpx handle}
}
\value{
Logical TRUE if catalog has data
}
\description{
Check if catalog has seed data
}
\keyword{internal}
