% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cli.R
\name{cli_progress}
\alias{cli_progress}
\title{Print a progress bar}
\usage{
cli_progress(current, total, width = 40)
}
\arguments{
\item{current}{Current progress}

\item{total}{Total items}

\item{width}{Bar width in characters}
}
\value{
Invisible NULL
}
\description{
Print a progress bar
}
\keyword{internal}
