% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroim2_utils.R
\name{concat_vecs}
\alias{concat_vecs}
\title{Concatenate multiple NeuroVec objects}
\usage{
concat_vecs(..., along = 4L)
}
\arguments{
\item{...}{NeuroVec objects to concatenate}

\item{along}{Dimension to concatenate along (default 4 = time)}
}
\value{
Combined NeuroVec with all time points.
}
\description{
Combines multiple NeuroVec objects along the time dimension.
Wrapper around neuroim2::concat() with validation.
}
\examples{
\dontrun{
h <- hcpx_ya()
run1 <- load_run(h, "100307_WM_LR_1")$dtseries
run2 <- load_run(h, "100307_WM_LR_2")$dtseries
combined <- concat_vecs(run1, run2)
}
}
