% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{derived_id}
\alias{derived_id}
\title{Generate derived product ID}
\usage{
derived_id(recipe, source_id, params = list())
}
\arguments{
\item{recipe}{Recipe name}

\item{source_id}{Source asset/task ID}

\item{params}{List of parameters}
}
\value{
Character derived ID
}
\description{
Generate derived product ID
}
\keyword{internal}
