% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_single}
\alias{download_single}
\title{Download a single file}
\usage{
download_single(h, remote_path, local_path, resume = TRUE)
}
\arguments{
\item{h}{hcpx handle}

\item{remote_path}{Remote path}

\item{local_path}{Local destination}

\item{resume}{Whether to resume partial downloads}
}
\value{
TRUE on success
}
\description{
Internal function to download one file using the appropriate backend.
}
\keyword{internal}
