% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends_rest.R
\name{ensure_session}
\alias{ensure_session}
\title{Ensure authenticated session exists}
\usage{
ensure_session(backend)
}
\arguments{
\item{backend}{REST backend object}
}
\value{
Backend with valid session
}
\description{
Ensure authenticated session exists
}
\keyword{internal}
