% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_paths.R
\name{extract_derivative_level}
\alias{extract_derivative_level}
\title{Extract derivative level from path}
\usage{
extract_derivative_level(path)
}
\arguments{
\item{path}{HCP remote path}
}
\value{
Character derivative level
}
\description{
Determines the processing stage based on path structure.
}
\keyword{internal}
