% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_paths.R
\name{extract_direction}
\alias{extract_direction}
\title{Extract direction (LR/RL) from path}
\usage{
extract_direction(path)
}
\arguments{
\item{path}{HCP remote path}
}
\value{
Character direction or NA
}
\description{
Extract direction (LR/RL) from path
}
\keyword{internal}
