% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_paths.R
\name{extract_file_type}
\alias{extract_file_type}
\title{Extract file type from path}
\usage{
extract_file_type(path)
}
\arguments{
\item{path}{HCP remote path}
}
\value{
Character file type
}
\description{
Determines the file type based on extension and path context.
}
\keyword{internal}
