% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_paths.R
\name{extract_kind}
\alias{extract_kind}
\title{Extract kind (struct, diff, rfmri, tfmri, meg) from path}
\usage{
extract_kind(path)
}
\arguments{
\item{path}{HCP remote path}
}
\value{
Character kind
}
\description{
Determines the data modality based on path patterns.
}
\keyword{internal}
