% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_paths.R
\name{extract_release}
\alias{extract_release}
\title{Extract release from path}
\usage{
extract_release(path)
}
\arguments{
\item{path}{HCP remote path}
}
\value{
Character release name or NA
}
\description{
Extract release from path
}
\keyword{internal}
