% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_paths.R
\name{extract_session}
\alias{extract_session}
\title{Extract session type from path}
\usage{
extract_session(path)
}
\arguments{
\item{path}{HCP remote path}
}
\value{
Character session type (3T, 7T, MEG) or NA
}
\description{
Extract session type from path
}
\keyword{internal}
