% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_paths.R
\name{extract_space}
\alias{extract_space}
\title{Extract space/coordinate system from path}
\usage{
extract_space(path)
}
\arguments{
\item{path}{HCP remote path}
}
\value{
Character space name
}
\description{
Extract space/coordinate system from path
}
\keyword{internal}
