% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_paths.R
\name{extract_subject_id}
\alias{extract_subject_id}
\title{Extract subject ID from path}
\usage{
extract_subject_id(path)
}
\arguments{
\item{path}{HCP remote path}
}
\value{
Character subject ID or NA
}
\description{
Extract subject ID from path
}
\keyword{internal}
