% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{get_access_tier}
\alias{get_access_tier}
\title{Check access tier of a variable}
\usage{
get_access_tier(h, var_name)
}
\arguments{
\item{h}{hcpx handle}

\item{var_name}{Variable name}
}
\value{
Character access tier ("open", "tier1", "tier2", or "unknown")
}
\description{
Check access tier of a variable
}
\keyword{internal}
