% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundles.R
\name{get_bundle}
\alias{get_bundle}
\title{Get bundle definition by name}
\usage{
get_bundle(bundle_name, h = NULL)
}
\arguments{
\item{bundle_name}{Name of the bundle}

\item{h}{hcpx handle (optional)}
}
\value{
List with bundle definition (where clause)
}
\description{
Get bundle definition by name
}
\keyword{internal}
