% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{get_completion_flag}
\alias{get_completion_flag}
\title{Get completion flag name for a task}
\usage{
get_completion_flag(task_name)
}
\arguments{
\item{task_name}{Canonical task name}
}
\value{
Completion flag column name or NA
}
\description{
Get completion flag name for a task
}
\keyword{internal}
