% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcpx.R
\name{get_con}
\alias{get_con}
\title{Get the database connection from an hcpx handle}
\usage{
get_con(h)
}
\arguments{
\item{h}{hcpx handle or hcpx_tbl}
}
\value{
DBI connection
}
\description{
Get the database connection from an hcpx handle
}
\keyword{internal}
