% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcpx.R
\name{get_hcpx}
\alias{get_hcpx}
\title{Extract hcpx handle from an hcpx_tbl}
\usage{
get_hcpx(x)
}
\arguments{
\item{x}{hcpx_tbl object}
}
\value{
hcpx handle
}
\description{
Extract hcpx handle from an hcpx_tbl
}
\keyword{internal}
