% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{get_recipe}
\alias{get_recipe}
\title{Get a recipe definition}
\usage{
get_recipe(name)
}
\arguments{
\item{name}{Recipe name}
}
\value{
Recipe definition list
}
\description{
Get a recipe definition
}
\keyword{internal}
