% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends_rest.R
\name{get_session_cookie}
\alias{get_session_cookie}
\title{Get current session cookie}
\usage{
get_session_cookie(backend)
}
\arguments{
\item{backend}{REST backend object}
}
\value{
Session cookie string or NULL
}
\description{
Get current session cookie
}
\keyword{internal}
