% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjects.R
\name{get_subject_ids}
\alias{get_subject_ids}
\title{Get subject IDs from an hcpx_tbl}
\usage{
get_subject_ids(x)
}
\arguments{
\item{x}{hcpx_tbl object}
}
\value{
Character vector of subject IDs
}
\description{
Extracts subject IDs from a subjects, tasks, or assets table.
}
\keyword{internal}
