% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{has_env_credentials}
\alias{has_env_credentials}
\title{Check if AWS credentials are in environment}
\usage{
has_env_credentials()
}
\value{
TRUE if credentials found
}
\description{
Check if AWS credentials are in environment
}
\keyword{internal}
