% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{has_keyring}
\alias{has_keyring}
\title{Check if keyring package is available}
\usage{
has_keyring()
}
\value{
TRUE if keyring is available
}
\description{
Check if keyring package is available
}
\keyword{internal}
