% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hash.R
\name{hash_md5}
\alias{hash_md5}
\title{Compute MD5 hash of a file}
\usage{
hash_md5(path)
}
\arguments{
\item{path}{File path}
}
\value{
Character MD5 hash
}
\description{
Compute MD5 hash of a file
}
\keyword{internal}
