% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{hcpx_auth_error}
\alias{hcpx_auth_error}
\title{Signal an authentication error}
\usage{
hcpx_auth_error(message, backend = NULL)
}
\arguments{
\item{message}{Error message}

\item{backend}{Backend type}
}
\description{
Signal an authentication error
}
\keyword{internal}
