% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends_rest.R
\name{hcpx_backend_rest}
\alias{hcpx_backend_rest}
\title{Create REST backend}
\usage{
hcpx_backend_rest(
  base_url = "https://db.humanconnectome.org",
  project = "HCP_1200"
)
}
\arguments{
\item{base_url}{Base URL for ConnectomeDB/XNAT}

\item{project}{XNAT project name (default: "HCP_1200")}
}
\value{
hcpx_backend_rest object
}
\description{
Create REST backend
}
\keyword{internal}
