% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcpx.R
\name{hcpx_cache}
\alias{hcpx_cache}
\title{Configure cache for an hcpx handle}
\usage{
hcpx_cache(
  h,
  root = NULL,
  max_size_gb = .DEFAULT_CACHE_MAX_GB,
  strategy = c("lru", "manual")
)
}
\arguments{
\item{h}{hcpx handle}

\item{root}{Cache root directory}

\item{max_size_gb}{Maximum cache size in GB}

\item{strategy}{One of "lru", "manual"}
}
\value{
Modified hcpx handle
}
\description{
Configure cache for an hcpx handle
}
\examples{
\dontrun{
h <- hcpx_ya()
h <- hcpx_cache(h, root = "/tmp/hcpx-cache", max_size_gb = 50)
}
}
