% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcpx.R
\name{hcpx_close}
\alias{hcpx_close}
\title{Close hcpx connection}
\usage{
hcpx_close(h)
}
\arguments{
\item{h}{hcpx handle}
}
\value{
Invisible NULL
}
\description{
Closes the database connection associated with an hcpx handle.
}
\examples{
\dontrun{
h <- hcpx_ya()
hcpx_close(h)
}
}
