% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{hcpx_download_error}
\alias{hcpx_download_error}
\title{Signal a download error}
\usage{
hcpx_download_error(message, path = NULL, status = NULL)
}
\arguments{
\item{message}{Error message}

\item{path}{Remote path}

\item{status}{HTTP status code (if applicable)}
}
\description{
Signal a download error
}
\keyword{internal}
