% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{hcpx_error}
\alias{hcpx_error}
\title{Create an hcpx error condition}
\usage{
hcpx_error(message, class = NULL, ...)
}
\arguments{
\item{message}{Error message}

\item{class}{Subclass name}

\item{...}{Additional condition data}
}
\value{
Condition object
}
\description{
Create an hcpx error condition
}
\keyword{internal}
