% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{hcpx_not_found}
\alias{hcpx_not_found}
\title{Signal that a resource was not found}
\usage{
hcpx_not_found(what, path = NULL)
}
\arguments{
\item{what}{Description of missing resource}

\item{path}{Path or identifier}
}
\description{
Signal that a resource was not found
}
\keyword{internal}
