% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{is_restricted}
\alias{is_restricted}
\title{Check if a variable is restricted}
\usage{
is_restricted(h, var_name)
}
\arguments{
\item{h}{hcpx handle}

\item{var_name}{Variable name}
}
\value{
Logical TRUE if restricted
}
\description{
Returns TRUE if the variable requires tier1 or tier2 access.
}
\keyword{internal}
