% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{keyring_get_credentials}
\alias{keyring_get_credentials}
\title{Get credentials from keyring}
\usage{
keyring_get_credentials(service = "hcpx")
}
\arguments{
\item{service}{Keyring service name}
}
\value{
List with access_key and secret_key, or NULL
}
\description{
Get credentials from keyring
}
\keyword{internal}
