% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{keyring_set_credentials}
\alias{keyring_set_credentials}
\title{Store credentials in keyring}
\usage{
keyring_set_credentials(service, access_key, secret_key)
}
\arguments{
\item{service}{Keyring service name}

\item{access_key}{AWS access key ID}

\item{secret_key}{AWS secret access key}
}
\value{
Invisible TRUE
}
\description{
Store credentials in keyring
}
\keyword{internal}
