% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{ledger_is_cached}
\alias{ledger_is_cached}
\title{Check which assets are cached}
\usage{
ledger_is_cached(h, asset_ids)
}
\arguments{
\item{h}{hcpx handle}

\item{asset_ids}{Character vector of asset IDs}
}
\value{
Logical vector (TRUE if cached)
}
\description{
Check which assets are cached
}
\keyword{internal}
