% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends_local.R
\name{local_path}
\alias{local_path}
\title{Build local path for a remote path}
\usage{
local_path(backend, remote_path)
}
\arguments{
\item{backend}{Local backend object}

\item{remote_path}{Remote path within the mirror}
}
\value{
Character path
}
\description{
Build local path for a remote path
}
\keyword{internal}
