% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{overview}
\alias{overview}
\alias{overview.hcpx}
\alias{overview.hcpx_tbl}
\title{Show what is available for a subject or cohort}
\usage{
overview(x)

\method{overview}{hcpx}(x)

\method{overview}{hcpx_tbl}(x)
}
\arguments{
\item{x}{hcpx_tbl object}
}
\value{
Invisible x, prints summary

Invisible x

Invisible x
}
\description{
Prints a beautiful summary of the data available in the catalog or
for a specific selection of subjects/tasks/assets.
}
\examples{
\dontrun{
h <- hcpx_ya()

# Overview of entire catalog
overview(h)

# Overview of female subjects
subjects(h, gender == "F") |> overview()

# Overview of WM task
tasks(h, "WM") |> overview()
}
}
