% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_paths.R
\name{parse_hcp_path}
\alias{parse_hcp_path}
\title{Parse HCP path to extract metadata}
\usage{
parse_hcp_path(path)
}
\arguments{
\item{path}{HCP remote path}
}
\value{
Named list with kind, task, direction, run, space, derivative_level, file_type
}
\description{
Parse HCP path to extract metadata
}
\keyword{internal}
