% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_paths.R
\name{parse_hcp_paths}
\alias{parse_hcp_paths}
\title{Vectorized path parsing}
\usage{
parse_hcp_paths(paths)
}
\arguments{
\item{paths}{Character vector of HCP remote paths}
}
\value{
Data frame with columns: kind, task, direction, run, space, derivative_level, file_type
}
\description{
Vectorized path parsing
}
\keyword{internal}
