% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_attach_handle}
\alias{plan_attach_handle}
\title{Attach an hcpx handle to a plan}
\usage{
plan_attach_handle(plan, h)
}
\arguments{
\item{plan}{hcpx_plan object}

\item{h}{hcpx handle}
}
\value{
Updated plan
}
\description{
Updates local paths and cache status based on the handle.
}
\keyword{internal}
