% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_cached}
\alias{plan_cached}
\title{Get files that are already cached}
\usage{
plan_cached(plan)
}
\arguments{
\item{plan}{hcpx_plan object}
}
\value{
Tibble of cached files
}
\description{
Get files that are already cached
}
\keyword{internal}
