% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_to_download}
\alias{plan_to_download}
\title{Get files that need downloading}
\usage{
plan_to_download(plan)
}
\arguments{
\item{plan}{hcpx_plan object}
}
\value{
Tibble of files to download
}
\description{
Get files that need downloading
}
\keyword{internal}
