% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{policy_summary}
\alias{policy_summary}
\title{Get summary of access tiers}
\usage{
policy_summary(h)
}
\arguments{
\item{h}{hcpx handle}
}
\value{
Tibble with access_tier and count
}
\description{
Returns count of variables in each access tier.
}
\examples{
\dontrun{
h <- hcpx_ya()
policy_summary(h)
}
}
