% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{read_manifest}
\alias{read_manifest}
\title{Read a plan manifest from disk}
\usage{
read_manifest(path, h = NULL)
}
\arguments{
\item{path}{File path to manifest JSON}

\item{h}{Optional hcpx handle to attach}
}
\value{
hcpx_plan object
}
\description{
Imports a manifest file and creates an hcpx_plan object. Note that the
plan will need an hcpx handle attached before downloading.
}
\examples{
\dontrun{
# Read manifest
plan <- read_manifest("my_plan.json")

# Attach handle and download
h <- hcpx_ya()
attr(plan, "hcpx") <- h
download(plan)
}
}
