% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends_rest.R
\name{rest_handle}
\alias{rest_handle}
\title{Create curl handle with session cookie}
\usage{
rest_handle(backend)
}
\arguments{
\item{backend}{REST backend object}
}
\value{
curl handle
}
\description{
Create curl handle with session cookie
}
\keyword{internal}
