% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends_aws.R
\name{s3_url}
\alias{s3_url}
\title{Build S3 URL for a remote path}
\usage{
s3_url(backend, remote_path)
}
\arguments{
\item{backend}{AWS backend object}

\item{remote_path}{Remote path within the bucket}
}
\value{
Character URL
}
\description{
Build S3 URL for a remote path
}
\keyword{internal}
