% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_build.R
\name{scan_s3_for_subjects}
\alias{scan_s3_for_subjects}
\title{Scan S3 bucket for assets for given subjects}
\usage{
scan_s3_for_subjects(h, subjects)
}
\arguments{
\item{h}{hcpx handle}

\item{subjects}{Character vector of subject IDs}
}
\value{
Number of assets added
}
\description{
Scan S3 bucket for assets for given subjects
}
\keyword{internal}
