% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_complete}
\alias{schema_complete}
\title{Check if all required tables exist}
\usage{
schema_complete(con)
}
\arguments{
\item{con}{DBI connection}
}
\value{
Logical TRUE if all tables exist
}
\description{
Check if all required tables exist
}
\keyword{internal}
