% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_create_bundles}
\alias{schema_create_bundles}
\title{Create bundles table}
\usage{
schema_create_bundles(con, engine)
}
\arguments{
\item{con}{DBI connection}

\item{engine}{Database engine identifier (reserved for future use)}
}
\value{
Invisible NULL
}
\description{
Create bundles table
}
\keyword{internal}
