% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_create_dictionary}
\alias{schema_create_dictionary}
\title{Create dictionary table}
\usage{
schema_create_dictionary(con, engine)
}
\arguments{
\item{con}{DBI connection}

\item{engine}{Database engine identifier (reserved for future use)}
}
\value{
Invisible NULL
}
\description{
Create dictionary table
}
\keyword{internal}
