% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_create_schema_version}
\alias{schema_create_schema_version}
\title{Create schema_version table}
\usage{
schema_create_schema_version(con, engine)
}
\arguments{
\item{con}{DBI connection}

\item{engine}{Database engine identifier (reserved for future use)}
}
\value{
Invisible NULL
}
\description{
Create schema_version table
}
\keyword{internal}
