% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_create_task_runs}
\alias{schema_create_task_runs}
\title{Create task_runs table}
\usage{
schema_create_task_runs(con, engine)
}
\arguments{
\item{con}{DBI connection}

\item{engine}{Database engine identifier (reserved for future use)}
}
\value{
Invisible NULL
}
\description{
Create task_runs table
}
\keyword{internal}
