% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_drop_all}
\alias{schema_drop_all}
\title{Drop all tables (for testing only)}
\usage{
schema_drop_all(con)
}
\arguments{
\item{con}{DBI connection}
}
\value{
Invisible NULL
}
\description{
Drop all tables (for testing only)
}
\keyword{internal}
