% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_needs_init}
\alias{schema_needs_init}
\title{Check if schema needs initialization}
\usage{
schema_needs_init(con)
}
\arguments{
\item{con}{DBI connection}
}
\value{
Logical TRUE if schema needs initialization
}
\description{
Check if schema needs initialization
}
\keyword{internal}
