% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_tables}
\alias{schema_tables}
\title{List all tables in the catalog database}
\usage{
schema_tables(con)
}
\arguments{
\item{con}{DBI connection}
}
\value{
Character vector of table names
}
\description{
List all tables in the catalog database
}
\keyword{internal}
