% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_version_get}
\alias{schema_version_get}
\title{Get current schema version}
\usage{
schema_version_get(con)
}
\arguments{
\item{con}{DBI connection}
}
\value{
Integer schema version, or NA if not initialized
}
\description{
Get current schema version
}
\keyword{internal}
