% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{total_size_bytes}
\alias{total_size_bytes}
\title{Get total size of assets in bytes}
\usage{
total_size_bytes(x)
}
\arguments{
\item{x}{hcpx_tbl assets table}
}
\value{
Numeric total size in bytes
}
\description{
Get total size of assets in bytes
}
\keyword{internal}
