% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundles.R
\name{validate_bundle}
\alias{validate_bundle}
\title{Validate bundle name}
\usage{
validate_bundle(bundle_name, h = NULL)
}
\arguments{
\item{bundle_name}{Bundle name to validate}

\item{h}{hcpx handle (optional)}
}
\value{
TRUE if valid, throws error if not
}
\description{
Validate bundle name
}
\keyword{internal}
