% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hash.R
\name{verify_checksum}
\alias{verify_checksum}
\title{Verify file checksum}
\usage{
verify_checksum(path, expected, algo = "md5")
}
\arguments{
\item{path}{File path}

\item{expected}{Expected checksum}

\item{algo}{Hash algorithm}
}
\value{
Logical TRUE if match
}
\description{
Verify file checksum
}
\keyword{internal}
